requirements.txt
pyOpenSSL
dnslib